requirements.txt
Flask
Flask_Cors
requests
ibm_watson_machine_learning
python-dotenv
